verifier: fix a race during reorgs
related: 41e088693de72d9111ce1ce21e42b5aa017627dd
If our guess of a txn getting confirmed at the same height in the new chain
as it was at in the old chain is incorrect, there is a race between the
verifier and the synchronizer. If the verifier wins, the exception will cause
us to disconnect.