wallet: fix channels backups in .backup files
This is a regression from https://github.com/spesmilo/electrum/commit/64a931f21e6f6699f192158e54ff8f2883682604,
which introduced "onchain_channel_backups", and renamed
the old "channel_backups" key to "imported_channel_backups".

The `save_backup` method was not changed to use the new "imported_channel_backups" key,
so the channel backups are in the backup file but they are ignored.